backend/requirements.txt
Flask
Flask-Cors
sentence-transformers
faiss-cpu
google-generativeai
numpy